refactor: 채팅 컨트롤러 권한 분리
- 채팅 내역 조회 및 채팅 참여자 조회는 토큰이 없어도 조회 가능하도록 한다.
